added a wrapper component around resultsList so that context can be passed successfully
